requirements.txt
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
annotated-types==0.6.0
anyio==4.3.0
apex==0.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.2.0
awscli==1.32.108
backcall==0.2.0
beautifulsoup4==4.12.3
botocore==1.34.108
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
cachetools==5.3.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi==1.16.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.6
crcmod==1.7
cryptography==41.0.7
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
-e git+ssh://git@github.com/CAPTAIN-WHU/iSAID_Devkit.git@a671fd53f748e3cb7489b86424f1dce266253461#egg=Detectron&subdirectory=preprocess/Detectron
distro==1.9.0
docutils==0.16
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.2.0
filelock==3.14.0
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2024.2.0
gdown==5.2.0
h11==0.14.0
h5py==3.7.0
httpcore==1.0.5
httpx==0.27.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
importlib-metadata==7.0.0
ipdb==0.13.13
ipython==8.12.3
jedi==0.18.2
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
jsonlines==4.0.0
kiwisolver==1.4.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
-e git+ssh://git@github.com/WTZ6666/Framework.git@c9a149122320371fd1e7e8dd7219777331c5385f#egg=maskrcnn_benchmark
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mmcv==1.6.0
mpmath==1.3.0
multidict==6.0.5
ninja==1.11.1
numpy==1.23.5
nvidia-ml-py==12.535.133
nvitop==1.3.2
openai==1.17.1
opencv-python==4.9.0.80
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.2
openxlab==0.0.31
ordered-set==4.1.0
oss2==2.17.0
overrides==7.3.1
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1701708255999/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1698245576425/work
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.0.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
requests==2.28.2
rich==13.4.2
rsa==4.7.2
s3transfer==0.10.1
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
synr==0.5.0
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tomli==2.0.1
torch==1.7.1+cu110
torch-geometric==2.0.4
torch-scatter==2.0.7
torch-sparse==0.6.9
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado==6.2
tqdm==4.65.2
traitlets==5.8.1
tvm==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.13
yacs==0.1.8
yapf==0.40.1
yarl==1.9.4
zipp==3.17.0
